+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to the "strike" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [Unreleased]
+### Changed
+- Whitespace chars are not striked by default.
+
+### Added
+- Settings to enable or disable whitespace strike (default is disabled).
+
 ## 1.0.1
 ### Fixed
 - Strike removal not working on toggle